Implement most of the remaining methods
I still need to properly implement editOrder() to allow replacement but I'd probably be best off refactoring the whole program to use a HashMap instead of an ArrayList.

I also need to implement confirmOrder() and that's just writing to a .csv file, so I can do that pretty easily later.